distmap: use a mid-way tight Cauchy bound estimate
Aberth-Ehrlich works by refining initial roots, similar to a more
elaborate Newton iterations algorithm. Choosing better initial estimates
leads to a faster convergence (meaning less iterations).

Cauchy bound define a space by defining the radius of a disk (complex
numbers are in 2D space) where all the roots of a polynomial should lie
within. Its "tighter" version (which adds the power part) is used here.

Since Aberth-Ehrlich is a refinement and not just a shrinking process,
we define and use an inner disk that has half the area of Cauchy bound
disk, so that we're more likely to start with initial guesses spread in
the "middle" of the roots; this is where the √2 comes from.