noise: fix 3d noise with negative coordinates
To reproduce the problem, use for example uv*2-1 instead of uv in
source_noise.frag, then check noise_{perlin,blocky} scenes from
tests/noise.py.

The fix makes sure we stay in the signed world as long as possible, and
only switch to unsigned for hashing.

The noise 1D implementation is currently left unchanged as the time
can't be negative, but it might make sense to change it as well there
eventually.