nodes: make sure the options are always accessed as read-only
With the exception of live change, node are not supposed to change the
node tree data.

ngli_node_get_data_ptr() is adjusted to accept the fallback data as
read-only since it's not modifying it, but we still cast instead of
returning it const because we don't want to impose a read-only usage to
the user.