glsl/path: rework glow effect
The glow has been reworked to make the whole shape glow, not just the
outline. The formula is also simpler and likely model better the light
decay with its use of the exponential. Thanks to this new formula, the
code could be made branchless (the division by 0 prevented that before).