text: make the outline position on the outer end by default
With text, the outline purpose is to make it readable on any background,
it's like adding a bold version of the same text behind with an opposite
color. Having it on the edge alters the shape of the character and is
very likely unwanted.

For path, the story is different so we keep it on the edge (centered) by
default.