distmap: make the SDF resolution agnostic
The scale defines the distance unit used in the SDF. Its value impacts
how far/fast an effect changes, for example an outline will be more or
less large, but does not impact the texture size; it can be seen as the
steepness of the slope of the SDF.

Before this change, its value was not the originally intended value: it
was calculated using pad which hasn't been computed yet. Moving pad
computation earlier has been considered, but the current chosen
heuristics for the scale is not exactly great anyway.

This change chooses a scale that is aligned with the resolution settings
(point size and DPI). This way, changing these settings will not have an
impact on the spread of the effects, only the quality.

Of course, a side effects is that some defaults and scenes are now a bit
off, so they will be tweaked a little in the following commits.